AI/requirements.txt
chromadb
google-generativeai
flask
flask-cors
langchain-community
langchain-google-genai
